issue-29 Move class sto separate file
