+++ src/main/kotlin/pullrequestfactory/Main.kt
@@ -14,10 +14,8 @@ import pullrequestfactory.domain.GithubPRFactory
 import pullrequestfactory.domain.PairingPartner
 import pullrequestfactory.domain.branches.BranchSyntaxValidator
 import pullrequestfactory.domain.pullrequests.PullRequestLastNotFinishedMarker
-import pullrequestfactory.io.programs.ProgramArgs
+import pullrequestfactory.io.clikt.CliktOpenPRsFactory
 import pullrequestfactory.io.programs.impl.FileAppProperties
-import pullrequestfactory.io.programs.impl.OpenPRProgramLastSessionFinished
-import pullrequestfactory.io.programs.impl.OpenPRsProgramLastSessionNotFinished
 import pullrequestfactory.io.repositories.GithubHttpBranchesRepos
 import pullrequestfactory.io.repositories.GithubHttpPullRequestsRepo
 import pullrequestfactory.io.repositories.KhttpClient
@@ -57,53 +55,16 @@ class Open : CliktCommand(help = "Open Pull Requests for Candidate") {
     ).defaultByName("regular")
 
     override fun run() {
-        val ui = ConsoleUI()
         val appProps = FileAppProperties("app.properties")
         val baseUrl = appProps.get_github_base_url()
         val repoPath = appProps.get_github_repository_path()
-        val repoUrl = baseUrl + repoPath
-        val program = when (val it = mode) {
-            is OpenModeInteractive -> {
-                echo("Opening Pull Requests for ${it.candidateFirstName} ${it.candidateFirstName}")
-                OpenPRsProgramLastSessionNotFinished(ui,
-                        repoUrl,
-                        KhttpClientStats(KhttpClient(it.githubAuthorizationToken)),
-                        Candidate(it.candidateFirstName, it.candidateLastName),
-                        create_pairing_partner("${it.pairingPartner1}-${it.pairingPartner2}-${it.pairingPartner3}-${it.pairingPartner4}-${it.pairingPartner5}-${it.pairingPartner6}-${it.pairingPartner7}"))
-            }
-            is OpenModeRegular -> {
-                val candidate = Candidate(it.candidate!!.split("-")[0], it.candidate!!.split("-")[1])
-                echo("Opening Pull Requests for ${candidate.firstName} ${candidate.lastName}")
-                val httpClient = KhttpClientStats(KhttpClient(it.githubToken!!))
-                val pairingPartner = create_pairing_partner(it.pairingPartner!!)
-                when (it.isLastIterationFinished) {
-                    true -> OpenPRProgramLastSessionFinished(ui,
-                            repoUrl,
-                            httpClient,
-                            candidate,
-                            pairingPartner)
-                    else -> OpenPRsProgramLastSessionNotFinished(ui,
-                            repoUrl,
-                            httpClient,
-                            candidate,
-                            pairingPartner)
-                }
-            }
-        }
+        val program = CliktOpenPRsFactory(ConsoleUI(), baseUrl + repoPath).create_for(mode)
         program.execute()
     }
+
 }
 
-private val ERROR_MSG_PAIRING_PARTNER = "Either option -p or pairing partner are missing or pairing partner have wrong format or is unknown"
-private fun create_pairing_partner(pairingPartner: String): List<PairingPartner> {
-    return pairingPartner.split("-").map { br ->
-        val pp = PairingPartner.value_of(br)
-        when (pp) {
-            null -> throw ProgramArgs.WrongPairingPartnerArgumentSyntax("$ERROR_MSG_PAIRING_PARTNER for given branch '$br'")
-            else -> pp
-        }
-    }
-}
+
 
 sealed class CloseMode(name: String) : OptionGroup(name)
